fix: refactor flow renderer for fast refresh
