+++ src/flow/index.ts
@@ -1,3 +1,11 @@
 export { FlowCanvas } from './FlowCanvas';
-export { renderFlow } from './renderFlow';
+export { RenderFlow } from './renderFlow';
+export {
+  createCustomNodeType,
+  createCustomEdgeType,
+  defaultTheme,
+  darkTheme,
+  applyTheme,
+} from './utils';
+export type { FlowTheme } from './utils';
 export type { FlowCanvasProps, FlowRenderer } from './types';
